refactor: move SchemaEditor code out of SchemaView
