+++ packages/graphql-toolbox/src/modules/SchemaView/SchemaEditor.tsx
@@ -19,15 +19,34 @@
 
 import { useContext, useEffect, useRef } from "react";
 
+import { autocompletion, closeBrackets, closeBracketsKeymap, completionKeymap } from "@codemirror/autocomplete";
+import { defaultKeymap, history, historyKeymap, indentWithTab } from "@codemirror/commands";
+import { bracketMatching, foldGutter, foldKeymap, indentOnInput } from "@codemirror/language";
+import { lintKeymap } from "@codemirror/lint";
+import { highlightSelectionMatches, searchKeymap } from "@codemirror/search";
+import { EditorState, StateEffect } from "@codemirror/state";
+import {
+    drawSelection,
+    dropCursor,
+    EditorView,
+    highlightActiveLine,
+    highlightSpecialChars,
+    keymap,
+    lineNumbers,
+} from "@codemirror/view";
 import { tokens } from "@neo4j-ndl/base";
 import { Button, IconButton, SmartTooltip } from "@neo4j-ndl/react";
 import { StarIconOutline } from "@neo4j-ndl/react/icons";
 import classNames from "classnames";
+import { graphql } from "cm6-graphql";
+import { dracula, tomorrow } from "thememirror";
 
 import { Extension, FileName } from "../../components/Filename";
-import { SCHEMA_EDITOR_INPUT } from "../../constants";
+import { DEFAULT_TYPE_DEFS, SCHEMA_EDITOR_INPUT } from "../../constants";
 import { Theme, ThemeContext } from "../../contexts/theme";
+import { useStore } from "../../store";
 import { handleEditorDisableState } from "../EditorView/utils";
+import { getSchemaForLintAndAutocompletion } from "./utils";
 
 export interface Props {
     loading: boolean;
@@ -37,6 +56,8 @@ export interface Props {
     introspect: () => Promise<void>;
     saveAsFavorite: () => void;
     onSubmit: () => void;
+    setEditorView: React.Dispatch<React.SetStateAction<EditorView | null>>;
+    editorView: EditorView | null;
 }
 
 export const SchemaEditor = ({
@@ -47,10 +68,73 @@ export const SchemaEditor = ({
     introspect,
     saveAsFavorite,
     onSubmit,
+    setEditorView,
+    editorView,
 }: Props) => {
     const theme = useContext(ThemeContext);
     const favoritesTooltipRef = useRef<HTMLButtonElement | null>(null);
     const introspectionTooltipRef = useRef<HTMLButtonElement | null>(null);
+    const storedTypeDefs = useStore.getState().typeDefinitions || DEFAULT_TYPE_DEFS;
+
+    const extensions = [
+        lineNumbers(),
+        highlightSpecialChars(),
+        highlightActiveLine(),
+        bracketMatching(),
+        closeBrackets(),
+        history(),
+        dropCursor(),
+        drawSelection(),
+        indentOnInput(),
+        autocompletion({ defaultKeymap: true, maxRenderedOptions: 5 }),
+        highlightSelectionMatches(),
+        EditorView.lineWrapping,
+        keymap.of([
+            indentWithTab,
+            ...closeBracketsKeymap,
+            ...defaultKeymap,
+            ...searchKeymap,
+            ...historyKeymap,
+            ...foldKeymap,
+            ...completionKeymap,
+            ...lintKeymap,
+        ]),
+        foldGutter({
+            closedText: "▶",
+            openText: "▼",
+        }),
+        graphql(getSchemaForLintAndAutocompletion()),
+        theme.theme === Theme.LIGHT ? tomorrow : dracula,
+    ];
+
+    useEffect(() => {
+        if (elementRef.current === null) {
+            return;
+        }
+
+        const state = EditorState.create({
+            doc: storedTypeDefs,
+            extensions,
+        });
+
+        const view = new EditorView({
+            state,
+            parent: elementRef.current,
+        });
+
+        setEditorView(view);
+
+        return () => {
+            view.destroy();
+            setEditorView(null);
+        };
+    }, [elementRef.current]);
+
+    useEffect(() => {
+        if (editorView) {
+            editorView.dispatch({ effects: StateEffect.reconfigure.of(extensions) });
+        }
+    }, [theme.theme, extensions]);
 
     useEffect(() => {
         handleEditorDisableState(elementRef.current, loading);
